src/referidos/requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.5.0
flask[async]==3.1.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet==3.2.3
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.6.0
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
py==1.11.0
PyDispatcher==2.0.6
pyparsing==3.0.9
pytest==8.4.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
psycopg2-binary
tomli==2.0.1
Werkzeug==3.1.3
